feat: Add a User Agent to cc-downloader
This commit adds a User Agent to cc-downloader so that its use can better be tracked.
It adds a refactor to the way the client is built so that the code is reused by both the download-paths and download functions.
It also adds 2 tests to check that the user agent is properly set.